Added Dac Client
